refactor(layering): evacuate cosmosdb/graphql/github/sharepoint providers into their connector crates
Continue the Phase-1 connector-only evacuations (docs/dev/crate_layering.md):
move each provider's data_components module into its existing connector-* crate
so the driver leaves runtime's prerequisite closure and data_components shrinks.

- cosmosdb -> connector-cosmosdb: azure_data_cosmos confined in restricted_deps
  (azure_core stays shared with runtime/runtime-secrets for Blob + Key Vault).
  data_components `cosmosdb` feature deleted; runtime pass-through -> [].
- graphql -> connector-graphql: rate_limit stays in data_components (used by
  runtime). graphql was never feature-gated, so only the module moves + the
  now-orphaned graphql-parser dep is dropped from data_components.
- github -> connector-github: takes a same-tier connector-graphql dep (github
  is built on the generic GraphQL provider). Also never feature-gated.
- sharepoint -> connector-sharepoint: graph-rs-sdk confined in restricted_deps;
  the sharepoint-mock-host test feature moves onto the connector; runtime's
  dev-dep integration test switches to connector_sharepoint::sharepoint.
  data_components `sharepoint`/`sharepoint-mock-host` features deleted; runtime
  pass-through -> [].

Layering guard green (130 crates, no upward edges); Cargo.lock regenerated.

elasticsearch is intentionally NOT evacuated: crates/search (shared-utility)
consumes data_components::elasticsearch::search_table, which depends on
query_table (also used by connector-elasticsearch), so moving the module to the
extension tier would create an illegal search -> connector-elasticsearch edge.
It needs a dedicated shared crate and is deferred (tracked with spark).
