+++ crates/data-connectors/connector-cosmosdb/Cargo.toml
@@ -7,15 +7,22 @@ license-file.workspace = true
 description = "Azure Cosmos DB data connector for Spice.ai runtime"
 
 [dependencies]
+arrow.workspace = true
 linkme.workspace = true
 async-trait.workspace = true
-data_components = { path = "../../data_components", features = ["cosmosdb"] }
+azure_core = "0.31.0"
+azure_data_cosmos = { version = "0.30.0", default-features = false, features = [
+    "hmac_rust",
+    "key_auth",
+] }
 datafusion.workspace = true
 datafusion-table-providers = { workspace = true }
+futures.workspace = true
 opentelemetry.workspace = true
 runtime = { path = "../../runtime" }
 runtime-api-types = { path = "../../runtime-api-types" }
 runtime-metrics = { path = "../../runtime-metrics" }
+serde_json.workspace = true
 snafu.workspace = true
 tokio.workspace = true
 tracing.workspace = true
