+++ crates/data-connectors/connector-cosmosdb/src/cosmosdb/client.rs
