+++ crates/data-connectors/connector-cosmosdb/src/cosmosdb/provider.rs
