+++ crates/data-connectors/connector-cosmosdb/src/cosmosdb/resilience.rs
