+++ crates/data-connectors/connector-github/Cargo.toml
@@ -28,6 +28,7 @@ arrow-schema.workspace = true
 arrow_tools = { path = "../../arrow_tools" }
 async-trait.workspace = true
 chrono.workspace = true
+connector-graphql = { path = "../connector-graphql" }
 data_components = { path = "../../data_components" }
 datafusion.workspace = true
 futures.workspace = true
@@ -40,12 +41,14 @@ runtime = { path = "../../runtime" }
 util = { path = "../../util" }
 runtime-rate-control = { path = "../../runtime-rate-control" }
 secrecy.workspace = true
+serde.workspace = true
 serde_json.workspace = true
 snafu.workspace = true
 token_provider = { path = "../../token_provider" }
 tokio.workspace = true
 tracing.workspace = true
 url.workspace = true
+zip = "8.3.0"
 
 [dev-dependencies]
 runtime-secrets = { path = "../../runtime-secrets", default-features = false }
