+++ crates/data-connectors/connector-github/src/workflow_runs.rs
@@ -14,8 +14,8 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
+use crate::github::{Error as GithubError, GithubRestClient};
 use async_trait::async_trait;
-use data_components::github::{Error as GithubError, GithubRestClient};
 use datafusion::{
     catalog::Session,
     common::{Column, Statistics},
@@ -274,7 +274,7 @@ impl WorkflowRunsTableProvider {
                 });
             } else if e
                 .downcast_ref::<GithubError>()
-                .is_some_and(data_components::github::Error::is_transient)
+                .is_some_and(crate::github::Error::is_transient)
             {
                 tracing::warn!(
                     "GitHub workflow runs provider initialization for {component} could not validate access because GitHub is temporarily unavailable: {e} The dataset will retry on the next query or refresh."
