+++ crates/data-connectors/connector-graphql/src/graphql/builder.rs
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-use crate::rate_limit::RateLimiter;
+use data_components::rate_limit::RateLimiter;
 
 use super::{Result, client::GraphQLClient, client::UnnestBehavior};
 use arrow::datatypes::SchemaRef;
