+++ crates/data-connectors/connector-sharepoint/Cargo.toml
@@ -7,20 +7,37 @@ license-file.workspace = true
 description = "SharePoint data connector for Spice.ai runtime"
 
 [dependencies]
+arrow.workspace = true
+async-stream.workspace = true
 async-trait.workspace = true
-data_components = { path = "../../data_components", features = ["sharepoint"] }
+base64.workspace = true
+bytes.workspace = true
+chrono.workspace = true
+data_components = { path = "../../data_components" }
 datafusion.workspace = true
 document_parse = { path = "../../document_parse" }
+futures.workspace = true
 graph-rs-sdk.workspace = true
+http.workspace = true
 linkme.workspace = true
+object_store = { workspace = true }
 paste.workspace = true
+percent-encoding.workspace = true
+reqwest.workspace = true
 runtime = { path = "../../runtime" }
 runtime-object-store = { path = "../../runtime-object-store" }
 secrecy.workspace = true
+serde.workspace = true
+serde_json.workspace = true
 snafu.workspace = true
 tokio.workspace = true
 tracing.workspace = true
 url.workspace = true
+
+[features]
+# Enables graph-rs-sdk's `test-util` feature so `GraphClient::use_test_endpoint`
+# accepts non-Microsoft hosts (e.g. localhost mocks). Intended for tests only.
+sharepoint-mock-host = ["graph-rs-sdk/test-util"]
 
 [lints]
 workspace = true
