+++ crates/data-connectors/connector-sharepoint/src/sharepoint/auth/mod.rs
