+++ crates/data-connectors/connector-sharepoint/src/sharepoint/auth/saml.rs
