+++ crates/data-connectors/connector-sharepoint/src/sharepoint/client.rs
