+++ crates/data-connectors/connector-sharepoint/src/sharepoint/error.rs
