+++ crates/data-connectors/connector-sharepoint/src/sharepoint/url.rs
