+++ crates/data_components/Cargo.toml
@@ -25,11 +25,6 @@ async-stream.workspace = true
 async-trait.workspace = true
 aws-credential-types.workspace = true
 aws-sdk-credential-bridge = { path = "../aws-sdk-credential-bridge" }
-azure_core = { version = "0.31.0", optional = true }
-azure_data_cosmos = { version = "0.30.0", default-features = false, features = [
-    "hmac_rust",
-    "key_auth",
-], optional = true }
 base64.workspace = true
 bb8 = { workspace = true, optional = true }
 bb8-oracle = { version = "0.3", features = ["chrono"], optional = true }
@@ -58,8 +53,6 @@ flight_client = { path = "../flight_client" }
 futures.workspace = true
 git2 = { workspace = true }
 globset.workspace = true
-graph-rs-sdk = { workspace = true, optional = true }
-graphql-parser.workspace = true
 hash-index = { path = "../hash-index", features = ["metrics"] }
 http.workspace = true
 httpdate.workspace = true
@@ -142,7 +135,6 @@ adbc = ["dep:adbc_core", "dep:adbc_driver_manager", "datafusion-table-providers/
 # upstream `impl SQLExecutor for SqlTable`, which is gated behind
 # `datafusion-table-providers/federation`.
 federation = ["datafusion-table-providers/federation"]
-cosmosdb = ["dep:azure_core", "dep:azure_data_cosmos"]
 databricks = [
     "delta_lake",
     "spark_connect",
@@ -183,10 +175,6 @@ postgres = [
 ]
 s3_vectors = ["dep:s3_vectors", "dep:s3_vectors_metadata_filter"]
 scylladb = ["dep:scylla", "datafusion-table-providers/federation"]
-sharepoint = ["dep:graph-rs-sdk"]
-# Enables graph-rs-sdk's `test-util` feature so `GraphClient::use_test_endpoint`
-# accepts non-Microsoft hosts (e.g. localhost mocks). Intended for tests only.
-sharepoint-mock-host = ["sharepoint", "graph-rs-sdk/test-util"]
 snowflake = [
     "dep:snowflake-api",
     "federation",
