+++ crates/runtime/Cargo.toml
@@ -295,7 +295,7 @@ aws-secrets-manager = ["dep:aws-sdk-sts", "runtime-secrets/aws-secrets-manager"]
 azure-keyvault = ["runtime-secrets/azure-keyvault"]
 hashicorp_vault = ["runtime-secrets/hashicorp_vault"]
 bedrock = []
-cosmosdb = ["data_components/cosmosdb"]
+cosmosdb = []
 cuda = ["llms/cuda"]
 databricks = ["data_components/databricks", "dep:duration-parse"]
 debezium = ["kafka", "data_components/debezium"]
@@ -378,7 +378,7 @@ s3_vectors = [
 ]
 scylladb = ["db_connection_pool/scylladb", "data_components/scylladb"]
 sftp = ["runtime-object-store/sftp"]
-sharepoint = ["data_components/sharepoint"]
+sharepoint = []
 smb = ["runtime-object-store/smb"]
 client_tls = []
 snapshots = ["runtime-acceleration/snapshots"]
